pyproject.toml
[tool.poetry]
name = "aris"
version = "0.1.0"
description = "Autonomous Research Intelligence System - AI-powered research assistant with document deduplication"
authors = ["ARIS Team"]
readme = "README.md"
packages = [{include = "aris", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# CLI and UI
click = "^8.1.7"
rich = "^13.7.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Storage and Database
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
chromadb = "^0.4.18"
gitpython = "^3.1.40"

# Security
keyring = "^24.3.0"
cryptography = "^41.0.7"

# Web and API
httpx = "^0.25.2"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
bleach = "^6.1.0"

# MCP Integration
mcp = "^0.9.0"

# Utilities
python-dateutil = "^2.8.2"
pyyaml = "^6.0.1"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.12.0"
ruff = "^0.1.8"
mypy = "^1.7.1"
bandit = "^1.7.5"

[tool.poetry.scripts]
aris = "aris.cli.main:cli"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "ANN", "S", "B", "A", "COM", "C4", "DTZ", "T10", "ISC", "ICN", "PIE", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/aris --cov-report=html --cov-report=term"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
